backend/requirements.txt
django>=2.2
djangorestframework>=3.10
django-ckeditor>=5.4.0
Pillow>=5.0.0
drf-yasg>=1.13.0
dj-database-url>=0.4.2
str2bool>=1.1
whitenoise>=4.1.2
gunicorn>=19.7.1
mysqlclient>=1.3.14
python-memcached>=1.59
django-cors-headers>=2.4.0
djangorestframework-camel-case>=1.0.3
django-filter>=2.0.0
django-admin-sortable>=2.1.7
django-allauth>=0.37.1
django-phonenumber-field>=2.1.0
inflection>=0.3.1
uritemplate>=3.0.0
dj-email-url>=0.1.0
django-admin-rangefilter>=0.3.8
phonenumbers>=8.9.16
django-import-export>=1.1.0
tablib>=0.13.0
ifaddr
djangorestframework-recursive
django-extensions
pydot
